feat: restructured cleanly new components
