+++ frontend/web/components/pages/environment-settings/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EnvironmentSettingsPage'
