+++ frontend/web/routes.js
@@ -17,7 +17,7 @@ import AccountSettingsPage from './components/pages/AccountSettingsPage'
 import NotFoundErrorPage from './components/pages/NotFoundErrorPage'
 import ProjectSettingsPage from './components/pages/project-settings'
 import PasswordResetPage from './components/pages/PasswordResetPage'
-import EnvironmentSettingsPage from './components/pages/EnvironmentSettingsPage'
+import EnvironmentSettingsPage from './components/pages/environment-settings'
 import InvitePage from './components/pages/InvitePage'
 import NotFoundPage from './components/pages/NotFoundPage'
 import ChangeRequestsPage from './components/pages/ChangeRequestsPage'
